requirements.txt
# filepath: d:\Blood_Donation_Workspace\vision_backend\requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
pytesseract
Pillow
PyMuPDF
pytest
pytest-cov
httpx
python-multipart
pydantic_settings
google-genai
python-dotenv
pymongo